refactor: 이미지 png -> webp로 교체
